+++ packages/ui/src/components/chat-ui/components/send-button.tsx
@@ -19,6 +19,7 @@ export const SendButton = (props: SendButtonProps) => {
 
   const button = (
     <IconButton
+      data-testid="send-message-button"
       size="sm"
       variant="solid"
       borderRadius="full"