+++ packages/ui/src/components/rich-text/shared/components/content-editable.tsx
@@ -1,5 +1,6 @@
 import { chakra, Flex } from "@chakra-ui/react";
 import { ContentEditable as LexicalContentEditable } from "@lexical/react/LexicalContentEditable";
+import { ScrollArea } from "@/components/scroll-area";
 
 const StyledEditable = chakra(LexicalContentEditable);
 
@@ -7,22 +8,32 @@ interface ContentEditableProps {
   onRef?: (elem: HTMLDivElement) => void;
   fullWidth?: boolean;
   padding?: string;
+  isRichMessage?: boolean;
 }
 
-export function ContentEditable({ onRef, fullWidth = false, padding }: ContentEditableProps) {
+export function ContentEditable({ onRef, fullWidth = false, padding, isRichMessage = false }: ContentEditableProps) {
   return (
-    <Flex className="rich-text" height="100%" overflow="auto" position="relative">
-      <Flex ref={onRef} width="100%" padding={padding} justifyContent={fullWidth ? "flex-start" : "center"}>
+    <ScrollArea
+      className="rich-text"
+      data-rich-message={isRichMessage ? "true" : undefined}
+      height="100%"
+      minH="0"
+      position="relative"
+      contentProps={{ width: "100%", minH: "100%" }}
+    >
+      <Flex ref={onRef} width="100%" minH="100%" padding={padding} justifyContent={fullWidth ? "flex-start" : "center"}>
         <StyledEditable
+          data-testid="content-editable"
           height="fit-content"
           minH="100%"
           width="100%"
           maxWidth={fullWidth ? "100%" : "720px"}
           textStyle="paragraph/M/regular"
+          fontSize="calc(var(--chakra-font-sizes-md) * var(--rich-text-font-scale))"
           position="relative"
           outline="none"
         />
       </Flex>
-    </Flex>
+    </ScrollArea>
   );
 }